possible bug fix by converting to double

+++ Framework/AnalysisModules/B2G-20-004/triggerlookup.cc
@@ -5,15 +5,15 @@ using namespace ROOT::VecOps;
 
 namespace analyzer {
     std::vector<float> TriggerLookup(float var, TH1D* TRP, TEfficiency* TEff ){
-        float Weight = 1;
-        float Weightup = 1;
-        float Weightdown = 1;
+        double Weight = 1;
+        double Weightup = 1;
+        double Weightdown = 1;
 
         std::vector<float> out;
 
         if (var < 4000.0){
             int bin0 = TRP->FindBin(var); 
-            float jetTriggerWeight = TEff->GetEfficiency(bin0);
+            double jetTriggerWeight = TEff->GetEfficiency(bin0);
             // Check that we're not in an empty bin in the fully efficient region
             if (jetTriggerWeight == 0){
                 if ((TEff->GetEfficiency(bin0-1) == 1.0) && (TEff->GetEfficiency(bin0+1) == 1.0)){
@@ -24,11 +24,11 @@ namespace analyzer {
             }
 
             Weight = jetTriggerWeight;
-            float deltaTriggerEff  = 0.5*(1.0-jetTriggerWeight);
-            float errorUp = TEff->GetEfficiencyErrorUp(bin0);
-            float errorDown = TEff->GetEfficiencyErrorLow(bin0);
-            float one = 1.0;
-            float zero = 0.0;
+            double deltaTriggerEff  = 0.5*(1.0-jetTriggerWeight);
+            double errorUp = TEff->GetEfficiencyErrorUp(bin0);
+            double errorDown = TEff->GetEfficiencyErrorLow(bin0);
+            double one = 1.0;
+            double zero = 0.0;
             Weightup = std::min(one,(jetTriggerWeight + sqrt(pow((deltaTriggerEff),2) + pow(errorUp,2) ))) ;
             Weightdown = std::max(zero,(jetTriggerWeight - sqrt(pow((deltaTriggerEff),2) + pow(errorDown,2) )));
